pyproject.toml
[tool.poetry]
name = "prs"
version = "2.5.5"
description = "Planning Referral System corporate application"
authors = ["Ashley Felton <ashley.felton@dbca.wa.gov.au>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.2.4"
psycopg2-binary = "2.8.6"
dbca-utils = "1.1.1"
django-dotenv = "1.4.2"
dj-database-url = "0.5.0"
gunicorn = "20.1.0"
dj-static = "0.0.6"
django-crispy-forms = "1.11.2"
django-reversion = "3.0.9"
django-tastypie = "0.14.3"
django-taggit = "1.4.0"
django-bootstrap-pagination = "1.7.1"
Unidecode = "1.2.0"
Pillow = "8.2.0"
python-magic = "0.4.22"
lxml = "4.6.3"
django-downloadview = "2.1.1"
openpyxl = "3.0.7"
xmltodict = "0.12.0"
djangorestframework = "3.12.4"
extract-msg = "0.28.7"
django-extensions = "3.1.3"
webtemplate-dbca = "1.3.0"
coverage = "4.5.1"
mixer = "6.1.3"
django-webtest = "1.9.7"
geojson = "2.5.0"
django-storages = {version = "1.11.1", extras = ["azure"]}

[tool.poetry.dev-dependencies]
ipython = "^7.23.0"
ipdb = "^0.13.2"

[build-system]
requires = ["poetry>=1.1.6"]
build-backend = "poetry.masonry.api"
